refactor 'moves' modules into their own sub-directories
