+++ app/src/lib/snake/moves/add/index.ts
@@ -1,5 +1,5 @@
-import { CardinalDirections, State } from "../types";
-import isValidMove from "./valid";
+import { CardinalDirections, State } from "../../types";
+import isValidMove from "../valid";
 
 const addMove = (state: State, move: CardinalDirections): State =>
   isValidMove(state, move)
